Find latest BlueSky model run from AirFire

Scans the directory of BluSky model output and returns the most recent modelRun string(s). If count > 1, the most recent count will be returned in low-hi order.

#' On 2019-10-11, available model identifiers include the following:

AK-12km

CANSAC-1.33km

CANSAC-4km

DRI1.33km-CMAQ

DRI4km-CMAQ

GFS-0.15deg-CanadaUSA-p25deg-68N

GFS-0.15deg

NAM-3km

NAM84-0.15deg

PNW-1.33km

PNW-4km

PNW1.33km-CMAQ

PNW4km-CMAQ

Users wanting the latest data will typically call this function first, followed by bluesky_load().

BlueSky output files are found in directories with the following structure: <baseUrl>/<modelName>/<modelRun>/<modelMode>/data/...

<baseUrl>/NAM84-0.08deg/2016050600/carryover/data/...
<baseUrl>/NAM84-0.08deg/2016050600/combined/data/...
<baseUrl>/NAM84-0.08deg/2016050600/forecast/data/...

bluesky_latestModelRun(
  modelName = NULL,
  count = 1,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  verbose = TRUE
)

Arguments

modelName
Model identifier.

count
Number of recent modelRun strings to return.

baseUrl
Base URL for BlueSky output.

verbose
If FALSE, suppress status messages (if any), and the progress bar.

Value

Timestamp of the latest model run.

Examples

# \donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

modelRun <- bluesky_latestModelRun(modelName = "PNW-4km")
filePath <- bluesky_download(modelName = "PNW-4km", modelRun = modelRun)

#> Auto-detected PNW-4km BlueSky Output Version: 2.0.1

bluesky_toCommonFormat(filePath)

#> [1] "/Users/jonathancallahan/Data/BlueSky/PNW-4km_2021031200_v2.nc"

bluesky_downloaded()

#>  [1] "CANSAC-1.33km_2019102900_v2.nc"                   
#>  [2] "CANSAC-1.33km_2021031112_v2.nc"                   
#>  [3] "CANSAC-1.33km_2021031200_v2.nc"                   
#>  [4] "CANSAC-4km_2019102700_v2.nc"                      
#>  [5] "CANSAC-4km_2019102900_v2.nc"                      
#>  [6] "CANSAC-4km_2021031112_v2.nc"                      
#>  [7] "CANSAC-4km_2021031200_v2.nc"                      
#>  [8] "GFS-0.15deg-CanadaUSA-p25deg-68N_2019102900_v2.nc"
#>  [9] "NAM-3km_2019102900.nc"                            
#> [10] "NAM84-0.15deg_2019102900_v2.nc"                   
#> [11] "PNW-1.33km_2019100900.nc"                         
#> [12] "PNW-1.33km_2020032500_v2.nc"                      
#> [13] "PNW-1.33km_2020091300_v2.nc"                      
#> [14] "PNW-1.33km_2021031112_v2.nc"                      
#> [15] "PNW-1.33km_2021031200_v2.nc"                      
#> [16] "PNW-4km_2019100900_v2.nc"                         
#> [17] "PNW-4km_2020032500_v2.nc"                         
#> [18] "PNW-4km_2020091300_v2.nc"                         
#> [19] "PNW-4km_2020091312_v2.nc"                         
#> [20] "PNW-4km_2020091400_v2.nc"                         
#> [21] "PNW-4km_2020091412_v2.nc"                         
#> [22] "PNW-4km_2020091500_v2.nc"                         
#> [23] "PNW-4km_2020091512_v2.nc"                         
#> [24] "PNW-4km_2020091600_v2.nc"                         
#> [25] "PNW-4km_2020091612_v2.nc"                         
#> [26] "PNW-4km_2020091700_v2.nc"                         
#> [27] "PNW-4km_2021010800_v2.nc"                         
#> [28] "PNW-4km_2021031112_v2.nc"                         
#> [29] "PNW-4km_2021031200_v2.nc"                         

# }