Load data for a single BlueSky model

This function encapsulates the process of downloading, and formatting necessary for BlueSky model outputs. Additionally, it checks already existing models and loads those if avaliable.

Loads BlueSky model output as a raster brick. The directory previously set with setModelDataDir() is searched for "v2" formatted model output. If data are not found locally, data are downloaded from baseUrl and converted.

The reurned dataBrick object is of class raster::RasterBrick and can be manipulated with appropriate functions from the raster package or any of the raster_~() functions provided by AirFireModeling.

On 2019-10-11, available model identifiers include the following:

AK-12km

CANSAC-1.33km

CANSAC-4km

DRI1.33km-CMAQ

DRI4km-CMAQ

GFS-0.15deg-CanadaUSA-p25deg-68N

GFS-0.15deg

NAM-3km

NAM84-0.15deg

PNW-1.33km

PNW-4km

PNW1.33km-CMAQ

PNW4km-CMAQ

bluesky_load(
  model = NULL,
  modelRun = NULL,
  modelType = "forecast",
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  localPath = NULL,
  xlim = NULL,
  ylim = NULL,
  clean = TRUE,
  verbose = TRUE
)

Arguments

model
Model identifier.

modelRun
Model initialization datestamp as "YYYYMMDDHH".

modelType
Subdirectory path containing BlueSky output, i.e. 'forcast'.

baseUrl
Base URL for BlueSky output.

localPath
Absolute path to a downloaded NetCDF file that is not found in `modelDataDir`.

xlim
A vector of coordinate longitude bounds.

ylim
A vector of coordinate latitude bounds.

clean
Logical specifying removal of original model data after conversion to "v2" format.

verbose
Logical to display messages.

Value

A raster package RasterBrick object.

Note

The regurned RasterBrick object will contain all grid cells within the area defined by xlim and ylim.

Examples

if (FALSE) {
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

raster <- bluesky_load(
  model = "PNW-4km",
  modelRun = 2019100900,
  xlim = c(-125, -115),
  ylim = c(42, 50)
)

raster_map(raster, index = 3)
}