Convert BlueSky NetCDF output to modernized NetCDF

Converts BlueSky model output from its original format to a more modernized NetCDF format with dimension axes for longitude, latitude, elevation and time. With default settings, output files renamed to ~base~_v2.nc.

bluesky_toCommonFormat(filePath = NULL, clean = TRUE)

Arguments

filePath
Absolute path of file to be converted.

clean
Logical specifying whether to remove the original netcdf file.

Value

Absolute path of the converted NetCDF file.

Note

Users will typically call bluesky_load() which in turn calls this function.

Examples

# \donttest{
library(AirFireModeling)
setModelDataDir('~/Data/BlueSky')

filePath <- bluesky_download(modelName = "PNW-4km", modelRun = 2020091300)

#> Auto-detected PNW-4km BlueSky Output Version: 2.0.0

bluesky_toCommonFormat(filePath)

#> [1] "/Users/jonathancallahan/Data/BlueSky/PNW-4km_2020091300_v2.nc"

bluesky_downloaded()

#>  [1] "CANSAC-1.33km_2019102900_v2.nc"                   
#>  [2] "CANSAC-1.33km_2021031112_v2.nc"                   
#>  [3] "CANSAC-1.33km_2021031200_v2.nc"                   
#>  [4] "CANSAC-4km_2019102700_v2.nc"                      
#>  [5] "CANSAC-4km_2019102900_v2.nc"                      
#>  [6] "CANSAC-4km_2021031112_v2.nc"                      
#>  [7] "CANSAC-4km_2021031200_v2.nc"                      
#>  [8] "GFS-0.15deg-CanadaUSA-p25deg-68N_2019102900_v2.nc"
#>  [9] "NAM-3km_2019102900.nc"                            
#> [10] "NAM84-0.15deg_2019102900_v2.nc"                   
#> [11] "PNW-1.33km_2019100900.nc"                         
#> [12] "PNW-1.33km_2020032500_v2.nc"                      
#> [13] "PNW-1.33km_2020091300_v2.nc"                      
#> [14] "PNW-1.33km_2021031112_v2.nc"                      
#> [15] "PNW-1.33km_2021031200_v2.nc"                      
#> [16] "PNW-4km_2019100900_v2.nc"                         
#> [17] "PNW-4km_2020032500_v2.nc"                         
#> [18] "PNW-4km_2020091300_v2.nc"                         
#> [19] "PNW-4km_2020091312_v2.nc"                         
#> [20] "PNW-4km_2020091400_v2.nc"                         
#> [21] "PNW-4km_2020091412_v2.nc"                         
#> [22] "PNW-4km_2020091500_v2.nc"                         
#> [23] "PNW-4km_2020091512_v2.nc"                         
#> [24] "PNW-4km_2020091600_v2.nc"                         
#> [25] "PNW-4km_2020091612_v2.nc"                         
#> [26] "PNW-4km_2020091700_v2.nc"                         
#> [27] "PNW-4km_2021010800_v2.nc"                         
#> [28] "PNW-4km_2021031112_v2.nc"                         
#> [29] "PNW-4km_2021031200_v2.nc"                         

# }