Correlation Map

Plots a map of the correlation of bluesky model data with a monitoring timeseries.

When colors=NULL, a color palette is given by colorRampPalette. Note: when using `plotter = ggplot`, zoom, colors and breaks are invalidated.

grid_correlationMap(bs_grid = NULL, ws_monitor = NULL,
  monitorID = NULL, parameter = "pm25", xlim = NULL, ylim = NULL,
  tlim = NULL, breaks = c(-1, -0.7, -0.5, -0.3, -0.1, 0.1, 0.3, 0.5,
  0.7, 1), colors = rev(RColorBrewer::brewer.pal(9, "RdBu")),
  zoom = 1.5, plotter = "base")

Arguments

bs_grid
bs_grid data list

ws_monitor
monitor data list

monitorID
monitor ID for correlation to model data

parameter
Data parameter to plot.

xlim
longitude range to be plotted

ylim
latitude range to be plotted

tlim
Time range tover which to calculation the correlation interpreted as UTC.

breaks
used by image() function to partition correlation values into different colors

colors
vector of colors to use

zoom
range extension as a fraction of the coordinate range

plotter
either "base" or "ggplot".

See also

Examples

if (FALSE) {
library(PWFSLSmoke)
library(AirFireModeling)

setModelDataDir("~/Data/Bluesky")

ws_monitor <- monitor_load(20191014, 20191016) %>%
  monitor_subset(monitorIDs = "lon_.120.591_lat_38.714_arb2.1008")
bs_grid <- bluesky_load(model = "CANSAC-1.33km", modelRun = 20191014)
grid_correlationMap(bs_grid, ws_monitor)
}