Subset a bs_grid with with a gridMask

This function subsets the bs_grid object and restructures the data into a ws_monitor that is compatible with PWFSLSmoke monitor_~ functions.

grid_subsetByMask(bs_grid = NULL, gridMask = NULL,
  countryCodes = c("CA", "US"))

Arguments

bs_grid
bs_grid object.

gridMask
Logical 2D matrix of the same dimensions as the grid data.

countryCodes
Vector of country codes covered by the grid.

Value

ws_monitor object where each unique monitor represents a single grid cell from bs_grid.

Note

The countryCodes parameter is provided because of the need to perform spatial searches for during the conversion of grid cell locations into ws_monitor objects. Each location must have correct information for countryCode, stateCode, timezone and these spatial searches can be time consuming. Specifying the countryCodes covered by the grid greatly speeds up these searches.

See also

Examples

if (FALSE) {
# Next three lines required for bs_grid to ws_monitor conversion
library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")
loadSpatialData("NaturalEarthAdm1")

# Now we can work as we normally would
setModelDataDir("~/Data/Bluesky")
bs <- bluesky_load(model = "PNW-1.33km", modelRun = 2019100900)
bs_monitors <- grid_subsetByDistance(bs, -121.017092, 46.901512,
                                     radius = 10000)

distanceMatrix <- grid_distance(bs, -121.017092, 46.901512)
radius <- 20000
gridMask <- distanceMatrix <= radius
bs_monitors <- grid_subsetByMask(bs, gridMask)
PWFSLSmoke::monitor_staticmap(bs_monitors, cex = .5, pch = 15)
}