Compare forecasts

Plots monitor data and overlays models of the bluesky forecasts.

For each model, a timeseries is created by gathering count grid cells within radius meters of the ws_monitor. The resulting set of timeseries are then collapsed into a single timeseries by applying the quantile() using the user specified probability prob.

monitor_forecastPlot(
  ws_monitor,
  modelName = NULL,
  modelMode = "forecast",
  baseUrl = "https://haze.airfire.org/bluesky-daily/output/standard",
  radius = 20,
  count = 9,
  prob = 0.5,
  verbose = TRUE,
  ...
)

Arguments

ws_monitor
ws_monitor object.

modelName
Model identifier(s).

modelMode
Subdirectory path containing BlueSky output, i.e. 'forcast'.

baseUrl
Base URL for BlueSky output.

radius
Distance (km) of radius from target location.

count
Number of grid cells within radius to return.

prob
Quantile probability used when plotting model data.

verbose
Logical to display messages.

...
Additional arguments passed to PWFSLSmoke::monitor_timeseriesPlot().

Examples

# \donttest{
library(PWFSLSmoke)

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> Loading required package: maps

#> Loading required package: MazamaCoreUtils

#> Loading required package: MazamaSpatialUtils

#> Loading required package: sp

initializeMazamaSpatialUtils()

library(AirFireModeling)
setModelDataDir('~/Data/Bluesky')

San_Pablo <-
  monitor_load(20191025, 20191029) %>%
  monitor_subset(monitorIDs = '060131004_01')

models <- bluesky_findModels(
  San_Pablo$meta$longitude,
  San_Pablo$meta$latitude
)

monitor_forecastPlot(
  San_Pablo,
  modelName = models
)

#> Converting raster to ws_monitor ...

#> Converting raster to ws_monitor ...

#> BlueSky Model Exists at: /Users/jonathancallahan/Data/Bluesky/NAM-3km_2019102900.nc

#> Converting raster to ws_monitor ...

#> Converting raster to ws_monitor ...

#> Converting raster to ws_monitor ...

#> Error: `vec_cast.POSIXct.POSIXct()` is implemented at C level.
#> This R function is purely indicative and should never be called.

# }