Load BlueSky NetCDF file and return contents as a bs_grid.

Reads in a "_V2" BlueSky output NetCDF file and ingests the data, returning it in a list of class bs_grid.

Data are searched for in the package internal dataDir which can be set with setModelDataDir(). If no data are found in the dataDir, and download = TRUE an attempt is made to download data.

If timesteps are specified, the returned bs_grid object will contain data associated with those timesteps. Data missing in the original model run will be replaced with NA values.

bluesky_load(dailyOutputDir = "standard", model = "PNW-1.33km",
  modelRun = NULL, subDir = "combined", parameter = "pm25",
  xlim = NULL, ylim = NULL, timesteps = NULL, download = TRUE,
  cleanup = TRUE, filePath = NULL,
  baseUrl = "https://haze.airfire.org/bluesky-daily/output",
  quiet = FALSE)

Arguments

dailyOutputDir
BlueSky web directory.

model
Model identifier.

modelRun
Date code as "YYYYMMDDHH" (integer or character).

subDir
Subdirectory path containing netcdf data.

parameter
Parameter name.

xlim
Optional longitude range.

ylim
Optional latitude range.

timesteps
Optional vector of timesteps to be loaded. Defaults to all available timesteps in the file.

download
Logical specifying whether to download and convert data if it is not found locally.

cleanup
Logical specifying whether to remove the original files.

filePath
Absolute path of the local bluesky .nc file (not "_V2.nc") for special occasions.

baseUrl
Base URL for BlueSky output.

quiet
If TRUE, suppress status messages (if any), and the progress bar.

Value

List of class bs_grid with the following elements:

lon -- 1D vector of longitudes (X)

lat -- 1D vector of latitudes (Y)

time -- 1D vector of POSIXct times (T)

data -- list of 3D arrays of data values (X-Y-T)

model -- character name of model used to generate the data

modelRun -- character date + hour when the model was run

deltaLon -- longitude axis grid cell spacing

deltaLat -- latitude axis grid cell spacing

BlueSky output files are found in directories with the following structure:

.../standard/NAM84-0.08deg/2016050600/carryover/data/...
.../standard/NAM84-0.08deg/2016050600/combined/data/...
.../standard/NAM84-0.08deg/2016050600/forecast/data/...

Details

#' On 2019-10-11, vailable model identifiers include the following:

AK-12km

CANSAC-1.33km

CANSAC-4km

DRI1.33km-CMAQ

DRI4km-CMAQ

GFS-0.15deg-CanadaUSA-p25deg-68N

GFS-0.15deg

NAM-3km

NAM84-0.15deg

PNW-1.33km

PNW-4km

PNW1.33km-CMAQ

PNW4km-CMAQ

See also

Examples

if (FALSE) {
setModelDataDir('~/Data/Bluesky')

xlim <- c(-118, -114)
ylim <- c(45, 48)
bs_grid <- bluesky_load(
  model = "PNW-1.33km",
  modelRun = 2019100900,
  xlim = xlim,
  ylim = ylim
)
grid_map(bs_grid, tMask = 1:24)
}